Code reformatted and optimized
