+++ src/test/java/com/softserveinc/edu/ita/enums/Roles.java
@@ -1,8 +1,10 @@
 package com.softserveinc.edu.ita.enums;
 
-
 import java.util.Random;
 
+/**
+ * Enum with roles names. Can generate random values
+ */
 public enum Roles {
     ADMINISTRATOR("Administrator"),
     CUSTOMER("Customer"),
@@ -21,7 +23,7 @@ public enum Roles {
         return this.role;
     }
 
-    public static Roles getRandomRole () {
+    public static Roles getRandomRole() {
         Random randomGenerator = new Random();
         return values()[randomGenerator.nextInt(values().length - 1)];
     }
