+++ src/test/java/com/softserveinc/edu/ita/page_object/PageObjectBase.java
@@ -15,6 +15,7 @@ public abstract class PageObjectBase {
     public String getElementText(By elementLocator) {
         return driver.findElement(elementLocator).getText();
     }
+
     //TODO consider using findElements()
     public boolean isElementDisplayed(By elementLocator) {
         return driver.findElement(elementLocator).isDisplayed();
@@ -33,7 +34,7 @@ public abstract class PageObjectBase {
         driver.findElement(elementLocator).click();
     }
 
-    public void clearElementText (By elementLocator) {
+    public void clearElementText(By elementLocator) {
         driver.findElement(elementLocator).clear();
     }
 }
