--- src/test/java/com/softserveinc/edu/ita/tests/TestSandBox.java
@@ -1,13 +1,0 @@
-package com.softserveinc.edu.ita.tests;
-
-import com.softserveinc.edu.ita.enums.Roles;
-
-/**
- * Created by i07016 on 10.06.2015.
- */
-public class TestSandBox {
-
-    public static void main(String[] args) {
-        Roles.getRandomRole();
-    }
-}
