+++ src/test/java/com/softserveinc/edu/ita/tests/UserInfoTest.java
@@ -5,17 +5,13 @@ import com.softserveinc.edu.ita.dataproviders.DataProviders;
 import com.softserveinc.edu.ita.locators.UserInfoPageLocators;
 import com.softserveinc.edu.ita.page_object.HomePage;
 import com.softserveinc.edu.ita.page_object.UserInfoPage;
-import org.apache.log4j.Logger;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-
 
 /**
  * Test information on User Info tab for any role User (Ticket IFAA-14)
  */
 public class UserInfoTest extends TestRunner {
-
-    Logger log = Logger.getLogger(UserInfoTest.class);
 
     @Test(dataProvider = "getAllRoles", dataProviderClass = DataProviders.class)
     public void testUserInfoTab(User user) {
