+++ src/test/java/com/softserveinc/edu/ita/utils/StringsGenerator.java
@@ -2,17 +2,17 @@ package com.softserveinc.edu.ita.utils;
 
 import java.io.IOException;
 import java.util.Random;
+
 import static com.softserveinc.edu.ita.utils.PropertyLoader.getProperty;
-/**
- *
- */
+
 public class StringsGenerator {
 
     /**
-     * Name generator. Can contain letters and numbers
+     * Random String generator
+     *
      * @param validSymbolsString set of symbols for generate string in valid_symbols.prop
-     * @param minimalLength minimal length of returned string
-     * @param maximalLength maximal length of returned string
+     * @param minimalLength      minimal length of returned string
+     * @param maximalLength      maximal length of returned string
      */
     public static String generateString(String validSymbolsString, int minimalLength, int maximalLength) {
 
@@ -26,7 +26,7 @@ public class StringsGenerator {
 
         try {
             String validSymbols = getProperty(validSymbolsString, SYMBOLS_FILENAME);
-            for (int i=0; i < generatedStringLength; i++) {
+            for (int i = 0; i < generatedStringLength; i++) {
                 generatedString.append(validSymbols
                         .charAt(randomGenerator.nextInt(validSymbols.length())));
             }
